Moving boilerplate code from *.data.js to helpers/data.js
