+++ source/demo/modules/imageversions/imageversions.data.js
@@ -1,15 +1,11 @@
 'use strict';
 
-var _ = require('lodash'),
-	requireNew = require('require-new'),
-	dataHelper = require('../../../../helpers/data.js'),
-	defaultData = requireNew('../../../data/default.data.js'),
-	data = _.merge(defaultData, {
-		meta: {
-			title: 'Demo: Image versions',
-			code: dataHelper.getTemplateCode('imageversions.hbs'),
-			documentation: dataHelper.getDocumentation('imageversions.md')
-		}
-	});
+var dataHelper = require('../../../../helpers/data.js');
 
-module.exports = data;
+module.exports = dataHelper.getExtendedData({
+	meta: {
+		title: 'Demo: Image versions',
+		code: dataHelper.getTemplateCode('imageversions.hbs'),
+		documentation: dataHelper.getDocumentation('imageversions.md')
+	}
+});
