+++ source/demo/modules/media/media.data.js
@@ -1,15 +1,11 @@
 'use strict';
 
-var _ = require('lodash'),
-	requireNew = require('require-new'),
-	dataHelper = require('../../../../helpers/data.js'),
-	defaultData = requireNew('../../../data/default.data.js'),
-	data = _.merge(defaultData, {
-		meta: {
-			title: 'Demo: Media demo',
-			jira: 'JIRA-3',
-			code: dataHelper.getTemplateCode('media.hbs')
-		}
-	});
+var dataHelper = require('../../../../helpers/data.js');
 
-module.exports = data;
+module.exports = dataHelper.getExtendedData({
+	meta: {
+		title: 'Demo: Media demo',
+		jira: 'JIRA-3',
+		code: dataHelper.getTemplateCode('media.hbs')
+	}
+});
