+++ source/demo/modules/skiplinks/skiplinks.data.js
@@ -1,23 +1,19 @@
 'use strict';
 
-var _ = require('lodash'),
-	requireNew = require('require-new'),
-	dataHelper = require('../../../../helpers/data.js'),
-	defaultData = requireNew('../../../data/default.data.js'),
-	data = _.merge(defaultData, {
-		meta: {
-			title: 'Demo: Skiplinks',
-			jira: 'JIRA-5',
-			code: dataHelper.getTemplateCode('skiplinks.hbs')
-		},
-		links: [
-			{
-				href: '#main',
-				accesskey: 1,
-				title: '[ALT + 1]',
-				label: 'Skip to content'
-			}
-		]
-	});
+var dataHelper = require('../../../../helpers/data.js');
 
-module.exports = data;
+module.exports = dataHelper.getExtendedData({
+	meta: {
+		title: 'Demo: Skiplinks',
+		jira: 'JIRA-5',
+		code: dataHelper.getTemplateCode('skiplinks.hbs')
+	},
+	links: [
+		{
+			href: '#main',
+			accesskey: 1,
+			title: '[ALT + 1]',
+			label: 'Skip to content'
+		}
+	]
+});
