+++ source/demo/modules/svgsprite/svgsprite.data.js
@@ -1,40 +1,16 @@
 'use strict';
 
-var _ = require('lodash'),
-	requireNew = require('require-new'),
-	glob = require('glob'),
-	path = require('path'),
-	spriteTask = require('../../../../gulp/media/svgsprite.js'),
-	dataHelper = require('../../../../helpers/data.js'),
-	defaultData = requireNew('../../../data/default.data.js'),
-	sprites = _.mapValues(spriteTask.taskConfig.src, function(globs) {
-		var files = [];
+var dataHelper = require('../../../../helpers/data.js');
 
-		globs.forEach(function(item) {
-
-			var paths = glob.sync(item);
-
-			paths = paths.map(function(file) {
-				return path.basename(file, path.extname(file));
-			});
-
-			files = files.concat(paths);
-		});
-
-		return files;
-	}),
-
-	data = _.merge(defaultData, {
-		meta: {
-			title: 'Demo: SVG icons',
-			jira: 'ESTATICO-212',
-			code: dataHelper.getTemplateCode('svgsprite.hbs'),
-			documentation: dataHelper.getDocumentation('svgsprite.md')
-		},
-		svgSprites: JSON.stringify(JSON.parse(defaultData.svgSprites || '[]').concat([
-			'/assets/media/svg/demo.svg'
-		])),
-		preview: sprites
-	});
-
-module.exports = data;
+module.exports = dataHelper.getExtendedData({
+	meta: {
+		title: 'Demo: SVG icons',
+		jira: 'ESTATICO-212',
+		code: dataHelper.getTemplateCode('svgsprite.hbs'),
+		documentation: dataHelper.getDocumentation('svgsprite.md')
+	},
+	svgSprites: JSON.stringify(JSON.parse(dataHelper.getDefaultData().svgSprites || '[]').concat([
+		'/assets/media/svg/demo.svg'
+	])),
+	preview: dataHelper.getSVGSprites()
+});
