+++ source/demo/pages/page/page.data.js
@@ -1,18 +1,15 @@
 'use strict';
 
-var _ = require('lodash'),
-	requireNew = require('require-new'),
-	defaultData = requireNew('../../../data/default.data.js'),
-	data = _.merge(defaultData, {
-		meta: {
-			title: 'Demo: 01 Page'
-		},
-		title: 'Page',
-		text: 'This page demonstrates the inclusion of a module.',
-		modules: {
-			skiplinks: requireNew('../../modules/skiplinks/skiplinks.data.js'),
-			teasers: requireNew('../../modules/teasers/teasers.data.js')
-		}
-	});
+var dataHelper = require('../../../../helpers/data.js');
 
-module.exports = data;
+module.exports = dataHelper.getExtendedData({
+	meta: {
+		title: 'Demo: 01 Page'
+	},
+	title: 'Page',
+	text: 'This page demonstrates the inclusion of a module.',
+	modules: {
+		skiplinks: dataHelper.tools.requireNew('../../modules/skiplinks/skiplinks.data.js'),
+		teasers: dataHelper.tools.requireNew('../../modules/teasers/teasers.data.js')
+	}
+});
