+++ source/pages/.scaffold/scaffold.data.js
@@ -1,17 +1,14 @@
 'use strict';
 
-var _ = require('lodash'),
-	requireNew = require('require-new'),
-	defaultData = requireNew('../../data/default.data.js'),
-	data = _.merge(defaultData, {
-		meta: {
-			title: '{{originalName}}'
-		},
-		title: '{{originalName}}',
-		text: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.',
-		modules: {
-			// xy: requireNew('../../modules/xy/xy.data.js')
-		}
-	});
+var dataHelper = require('../../../helpers/data.js');
 
-module.exports = data;
+module.exports = dataHelper.getExtendedData({
+	meta: {
+		title: '{{originalName}}'
+	},
+	title: '{{originalName}}',
+	text: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.',
+	modules: {
+		// xy: requireNew('../../modules/xy/xy.data.js')
+	}
+});
